health_call_agent/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
langgraph>=0.0.40
langchain-google-genai>=2.0.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
